// src/interfaces/physical_attributes.interface.ts
export interface PhysicalAttributesFormData {
  height: string;
  weight: string;
  body_type: string;
  complexion: string;
  physical_status: string;
  blood_group: string;
  eye_color: string;
  hair_color: string;
  hair_type: string;
  hair_length: string;
  skin_tone: string;
  fitness_level: string;
  disability: string;
  disability_details: string;
  spectacles: string;
  lens_usage: string;
  beard_style: string;
  tattoo: string;
  physique: string;
  shoe_size: string;
  dress_size: string;
  health_condition: string;
  medical_conditions: string;
  genetic_disorders: string;
  appearance_notes: string;
}

export const initialPhysicalAttributesFormData: PhysicalAttributesFormData = {
  height: "",
  weight: "",
  body_type: "",
  complexion: "",
  physical_status: "",
  blood_group: "",
  eye_color: "",
  hair_color: "",
  hair_type: "",
  hair_length: "",
  skin_tone: "",
  fitness_level: "",
  disability: "",
  disability_details: "",
  spectacles: "",
  lens_usage: "",
  beard_style: "",
  tattoo: "",
  physique: "",
  shoe_size: "",
  dress_size: "",
  health_condition: "",
  medical_conditions: "",
  genetic_disorders: "",
  appearance_notes: "",
};